api: Don't allocate arrays in api call function
When having channels with high number of elements, most of the time
spent during execution is for allocating the samples and timestamps
arrays.

Fix this by moving the allocation from the api call funtions up outside.
Now, we only allocate if the block size changes, which happens for the
first and the last block.